branches/main: cisstMultiTask, minor fixes to compile and reduce warnings
git-svn-id: https://svn.lcsr.jhu.edu/cisst/branches/main@767 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ tests/cisstMultiTask/mtsTaskTest.cpp
@@ -41,9 +41,9 @@ mtsTaskTestTask::mtsTaskTestTask(const std::string & collectorName,
 //-----------------------------------------------------------------------------
 void mtsTaskTest::TestGetStateVectorID(void)
 {
-	const int default_column_count = 3;	// Toc, Tic, Period
-	const int user_column_count = 2;	// Data1, Data2
-	const int total_column_count = default_column_count + user_column_count;
+	const size_t default_column_count = 3;	// Toc, Tic, Period
+	const size_t user_column_count = 2;	// Data1, Data2
+	const size_t total_column_count = default_column_count + user_column_count;
 	
 	const std::string names[total_column_count] = { 
 		// added by default
